docs: add more Jumpy API documentation
- improved main page and several modules.
- also move state enums into their own module for readability.
