+++ src/prelude.rs
@@ -1,6 +1,6 @@
 pub use crate::{
-    assets::*, audio::*, camera::*, config::*, debug::*, input::*, loading::*, localization::*,
-    metadata::*, platform::*, session::*, ui::*, utils::*, *,
+    assets::*, audio::*, bevy_states::*, camera::*, config::*, debug::*, input::*, loading::*,
+    localization::*, metadata::*, platform::*, session::*, ui::*, utils::*, *,
 };
 pub use anyhow::Context;
 pub use jumpy_core::bevy_prelude::*;
